Only allow the anti lockout rule to operate on LAN
This does not fix all the problems with the anti lockout rule logic, but
it makes the bug less catastrophic:
If the bug stroke, you either had open ports on WAN (or on unwanted
interfaces) or the anti lockout rule simply did not work.
Now if the bug hits, the anti lockout rule will simply not work.

For details see #7372